added header button to copy scene uuid
