+++ classes/macroBrowser.js
@@ -1,19 +1,20 @@
 import SceneMacrosData from "./sceneMacrosData.js"
+import { copyToClipboard } from "./utils.js"
 const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api
 
-export default class MacroBroswerV2 extends HandlebarsApplicationMixin(ApplicationV2) {
+export default class MacroBrowserV2 extends HandlebarsApplicationMixin(ApplicationV2) {
 
     static DEFAULT_OPTIONS = {
         actions: {
-            comment: MacroBroswerV2.comment,
-            copyUuid: MacroBroswerV2.#copyUuid,
-            edit: MacroBroswerV2.#editMacro,
-            execute: MacroBroswerV2.#executeMacro,
-            unlink: MacroBroswerV2.#unlinkMacro
+            comment: MacroBrowserV2.comment,
+            copyUuid: MacroBrowserV2.#copyUuid,
+            edit: MacroBrowserV2.#editMacro,
+            execute: MacroBrowserV2.#executeMacro,
+            unlink: MacroBrowserV2.#unlinkMacro
         },
         classes: ['standard-form'],
         form: {
-            handler: MacroBroswerV2.#onSubmit,
+            handler: MacroBrowserV2.#onSubmit,
         },
         id: 'macro-browser_{id}',
         position: {
@@ -83,37 +84,7 @@ export default class MacroBroswerV2 extends HandlebarsApplicationMixin(Applicati
     // COPIES A UUID TO CLIPBOARD 
     static #copyUuid(event, target) {
         const text = target.dataset.uuid
-        if (typeof text === 'string') {
-            try {
-                if (navigator.clipboard && window.isSecureContext) {
-                    navigator.clipboard.writeText(text)
-                } else {
-                    // fallback in case not in HTTPS - the jankiest method to copy text know to man, like wtf
-                    const copyText = document.createElement('textarea')
-                    copyText.value = text
-                    copyText.style.position = 'absolute';
-                    copyText.style.left = '-99999px'
-
-                    document.body.prepend(copyText)
-                    copyText.select()
-
-                    try {
-                        document.execCommand('copy')
-                    } catch (err) {
-                        console.error(err)
-                    } finally {
-                        copyText.remove()
-                    }
-                }
-                ui.notifications.info(game.i18n.localize("SCENE_MACROS.macro-browser.copied"))
-            } catch (err) {
-                console.error(err)
-                ui.notifications.error(game.i18n.localize("SCENE_MACROS.macro-browser.copy-failed"))
-            }
-        } else {
-            ui.notifications.error(game.i18n.localize("SCENE_MACROS.macro-browser.copy-failed"))
-            console.error('TypeError: element to copy is not a string', text, this)
-        }
+        copyToClipboard(text)
     }
 
     // OPEN EDITOR FOR MACRO
@@ -163,4 +134,15 @@ export default class MacroBroswerV2 extends HandlebarsApplicationMixin(Applicati
     }
 
 }
+
+Hooks.on('getHeaderControlsMacroBrowserV2', (application, controls) => {
+    controls.push({
+        icon: 'fas fa-passport',
+        label: game.i18n.localize("SCENE_MACROS.macro-browser.copy-scene"),
+        onClick: () => {
+            copyToClipboard(application.document.uuid)
+        },
+        ownership: CONST.DOCUMENT_OWNERSHIP_LEVELS['OWNER']
+    })
+})
